### 2023-02-02 #572 @lyh970817
qualtdict: Generating Variable Dictionaries and Labelled Data Exports of Qualtrics Surveys
Submitting Author Name: Yuhao Lin
Submitting Author Github Handle: @lyh970817
Repository:  https://github.com/lyh970817/qualtdict
Version submitted: 0.0.0.9000
Submission type: Standard

Version accepted: TBD
---

```
Package: qualtdict
Title: Generating Variable Dictionaries and Labelled Data Exports of Qualtrics
    Surveys
Version: 0.0.0.9000
    person("Yuhao", "Lin", , "yuhao.lin@kcl.ac.uk", role = c("aut", "cre"),
Description: Provides functions that generate variable dictionaries from
    'Qualtrics' <https://www.qualtrics.com/about/> surveys and labelled
    survey data based on the dictionary. This package is built upon the R
    package 'qualtRics' <https://github.com/ropensci/qualtRics/> which
    provides access to 'Qualtrics' survey data and metadata via the 'Qualtrics' API
    <https://api.qualtrics.com/>.
    crul,
    glue,
    haven,
    openNLP,
    qualtRics,
    sjlabelled,
    slowraker,
    SnowballC,
    stringi,
    xml2
    covr,
    vcr (>= 0.6.0)
    knitr
Config/testthat/start-first: dict_generate, dict_validate, get_survey_data
```

[Qualtrics](https://www.qualtrics.com/) is an online survey and data collection software platform. While the [qualtRics](https://github.com/ropensci/qualtRics) R package implements data retrieval from the Qualtrics platform, this package 'qualtdict' further processes its output to generate variable dictionaries and labelled data designed to be used for data analyses directly.

The target audience is those who use the Qualtrics survey platform to collect data. This package generates variable dictionaries and labelled data designed to be used for data analyses directly.

No, but there is the similar [qualtRics](https://github.com/ropensci/qualtRics) R package that retrieves a broader range of data from Qualtrics than this package utilises. The output formats from qualtRics are much less user-friendly, for example, it retrieves survey metadata in a nested-list, json-like format, while this package rearranges essential parts of this metadata (retrieved using quatRics) into a publishable variable dictionary in a table format that can be visually inspected in, for example, excel.

Yes.
