+++ tests/testthat/_snaps/dict_validate.md
@@ -10763,17 +10763,17 @@
                 {
                   "type": "character",
                   "attributes": {},
-                  "value": ["QID13_1", "QID14_1", "QID20_1", "QID21_1", "QID22_1", "QID33_1", "QID36_1", "QID38_1", "QID40_1", "QID4_1"]
+                  "value": ["QID13_1", "QID14_1", "QID20_1", "QID21_1", "QID22_1", "QID33_1", "QID38_1", "QID40_1"]
                 },
                 {
                   "type": "character",
                   "attributes": {},
-                  "value": ["QID13_2", "QID14_2", "QID20_2", "QID21_2", "QID22_2", "QID33_2", "QID36_2", "QID38_2", "QID40_2", "QID4_2"]
+                  "value": ["QID13_2", "QID14_2", "QID20_2", "QID21_2", "QID22_2", "QID33_2", "QID38_2", "QID40_2"]
                 },
                 {
                   "type": "character",
                   "attributes": {},
-                  "value": ["QID13_3", "QID14_3", "QID20_3", "QID21_3", "QID22_3", "QID33_3", "QID36_3", "QID38_3", "QID40_3", "QID4_3"]
+                  "value": ["QID13_3", "QID14_3", "QID20_3", "QID21_3", "QID22_3", "QID33_3", "QID38_3", "QID40_3"]
                 },
                 {
                   "type": "character",
@@ -10868,7 +10868,7 @@
                 {
                   "type": "character",
                   "attributes": {},
-                  "value": ["QID23_FILE_ID", "QID23_FILE_NAME", "QID23_FILE_SIZE", "QID23_FILE_TYPE", "QID24", "QID26_FILE_ID", "QID26_FILE_NAME", "QID26_FILE_SIZE", "QID26_FILE_TYPE", "QID2_TEXT", "QID37", "QID49", "QID50"]
+                  "value": ["QID23_FILE_ID", "QID23_FILE_NAME", "QID23_FILE_SIZE", "QID23_FILE_TYPE", "QID24", "QID26_FILE_ID", "QID26_FILE_NAME", "QID26_FILE_SIZE", "QID26_FILE_TYPE", "QID2_TEXT", "QID37", "QID49", "QID50", "QID52", "QID53"]
                 },
                 {
                   "type": "character",
@@ -11523,7 +11523,7 @@
                 {
                   "type": "character",
                   "attributes": {},
-                  "value": ["QID32", "QID39", "QID41", "QID42", "QID43", "x1_QID31", "x2_QID31", "x3_QID31"]
+                  "value": ["QID32", "QID36", "QID39", "QID4", "QID41", "QID42", "QID43", "x1_QID31", "x2_QID31", "x3_QID31"]
                 },
                 {
                   "type": "character",
