+++ tests/fixtures/dict_generate.yml
@@ -17,16 +17,16 @@ http_interactions:
       reason: OK
       message: 'Success: (200) OK'
     headers:
-      request-time: '608'
+      request-time: '661'
       x-frame-options: DENY
       x-xss-protection: 1; mode=block
       x-permitted-cross-domain-policies: master-only
       content-type: application/json
-      content-length: '65224'
+      content-length: '65779'
       referrer-policy: strict-origin-when-cross-origin
       x-content-type-options: nosniff
       permissions-policy: camera=(), geolocation=(), microphone=()
-      date: Sun, 29 Jan 2023 08:56:53 GMT
+      date: Fri, 03 Feb 2023 13:32:28 GMT
       content-security-policy-report-only: frame-ancestors 'self' *.qualtrics.com
         *.my.salesforce.com *.visualforce.com *.visual.force.com *.lightning.force.com;
         report-uri https://sjc1.qualtrics.com/csp-report
@@ -34,7 +34,7 @@ http_interactions:
     body:
       encoding: ''
       file: no
-      string: '{"result":{"id":"SV_0AQg1pFepA0V2d0","name":"Test","ownerId":"UR_cYCFQWCwckGHpf7","organizationId":"kcliop","isActive":false,"creationDate":"2022-02-08T09:36:16Z","lastModifiedDate":"2023-01-29T07:32:10Z","expiration":{"startDate":null,"endDate":null},"questions":{"QID1":{"questionType":{"type":"MC","selector":"SAVR","subSelector":"TX"},"questionText":"Multiple
+      string: '{"result":{"id":"SV_0AQg1pFepA0V2d0","name":"Test","ownerId":"UR_cYCFQWCwckGHpf7","organizationId":"kcliop","isActive":false,"creationDate":"2022-02-08T09:36:16Z","lastModifiedDate":"2023-02-03T13:15:46Z","expiration":{"startDate":null,"endDate":null},"questions":{"QID1":{"questionType":{"type":"MC","selector":"SAVR","subSelector":"TX"},"questionText":"Multiple
         choice one answer","questionLabel":null,"validation":{"doesForceResponse":false},"questionName":"Q1","choices":{"1":{"recode":"1","description":"Click
         to write Choice 1","choiceText":"Click to write Choice 1","imageDescription":null,"variableName":null,"analyze":true},"2":{"recode":"2","description":"Click
         to write Choice 2","choiceText":"Click to write Choice 2","imageDescription":null,"variableName":null,"analyze":true},"3":{"recode":"1","description":"Click
@@ -75,9 +75,11 @@ http_interactions:
         to write Choice 1","choiceText":"Click to write Choice 1","imageDescription":null,"variableName":null,"analyze":true},"2":{"recode":"2","description":"Click
         to write Choice 2","choiceText":"Click to write Choice 2","imageDescription":null,"variableName":null,"analyze":true},"3":{"recode":"3","description":"Click
[…]
