+++ .pre-commit-config.yaml
@@ -57,6 +57,7 @@ repos:
           (.*/|)WORDLIST|
           \.github/workflows/.*|
           data/.*|
+          tests/.*|
           )$
     -   id: lintr
     -   id: readme-rmd-rendered
@@ -72,7 +73,7 @@ repos:
     -   id: file-contents-sorter
         files: '^\.Rbuildignore$'
     -   id: end-of-file-fixer
-        exclude: '\.Rd'
+        exclude: '\.Rd|tests/.*'
 -   repo: https://github.com/pre-commit-ci/pre-commit-ci-config
     rev: v1.5.1
     hooks:
