+++ tests/fixtures/get_survey_data.yml
@@ -17,7 +17,7 @@ http_interactions:
       reason: OK
       message: 'Success: (200) OK'
     headers:
-      request-time: '339'
+      request-time: '503'
       x-frame-options: DENY
       x-xss-protection: 1; mode=block
       x-permitted-cross-domain-policies: master-only
@@ -26,7 +26,7 @@ http_interactions:
       referrer-policy: strict-origin-when-cross-origin
       x-content-type-options: nosniff
       permissions-policy: camera=(), geolocation=(), microphone=()
-      date: Sun, 29 Jan 2023 08:57:14 GMT
+      date: Sat, 04 Feb 2023 04:50:52 GMT
       content-security-policy-report-only: frame-ancestors 'self' *.qualtrics.com
         *.my.salesforce.com *.visualforce.com *.visual.force.com *.lightning.force.com;
         report-uri https://sjc1.qualtrics.com/csp-report
@@ -34,13 +34,13 @@ http_interactions:
     body:
       encoding: ''
       file: no
-      string: '{"result":{"progressId":"ES_4JxfCXwJk10Gd3U","percentComplete":0.0,"status":"inProgress"},"meta":{"requestId":"36de8569-e029-4452-9982-6036b528708e","httpStatus":"200
+      string: '{"result":{"progressId":"ES_eWfbiGTnNZPRzsW","percentComplete":0.0,"status":"inProgress"},"meta":{"requestId":"bc77491c-bb1a-48ea-85fb-7eceedc04a0e","httpStatus":"200
         - OK"}}'
-  recorded_at: 2023-01-29 08:57:19 GMT
+  recorded_at: 2023-02-04 04:51:00 GMT
   recorded_with: vcr/1.2.0, webmockr/0.8.2
 - request:
     method: get
-    uri: https://QUALTRICS_BASE_URL/API/v3/surveys/SV_0AQg1pFepA0V2d0/export-responses/ES_4JxfCXwJk10Gd3U
+    uri: https://QUALTRICS_BASE_URL/API/v3/surveys/SV_0AQg1pFepA0V2d0/export-responses/ES_eWfbiGTnNZPRzsW
     body:
       encoding: ''
       string: ''
@@ -56,7 +56,7 @@ http_interactions:
       reason: OK
       message: 'Success: (200) OK'
     headers:
-      request-time: '204'
+      request-time: '351'
       x-frame-options: DENY
       x-xss-protection: 1; mode=block
       x-permitted-cross-domain-policies: master-only
@@ -65,7 +65,7 @@ http_interactions:
       referrer-policy: strict-origin-when-cross-origin
       x-content-type-options: nosniff
       permissions-policy: camera=(), geolocation=(), microphone=()
-      date: Sun, 29 Jan 2023 08:57:15 GMT
+      date: Sat, 04 Feb 2023 04:50:53 GMT
       content-security-policy-report-only: frame-ancestors 'self' *.qualtrics.com
         *.my.salesforce.com *.visualforce.com *.visual.force.com *.lightning.force.com;
         report-uri https://sjc1.qualtrics.com/csp-report
@@ -73,13 +73,13 @@ http_interactions:
[…]
