+++ R/qid_recode.R
@@ -234,7 +234,11 @@ qid_recode <- function(qid,
           SACH = item_or_level_qid,
           SACCOL = item_or_level_qid
         ),
-        TE = list(Short = suf_item_suf_level_qid),
+        TE = list(
+          Short = suf_item_suf_level_qid,
+          Medium = suf_item_suf_level_qid,
+          Long = suf_item_suf_level_qid
+        ),
         Profile = list(
           SingleAnswer = suf_item_rep_level_qid,
           DL = suf_item_rep_level_qid
@@ -271,13 +275,13 @@ qid_recode <- function(qid,
     HL = list(Text = suf_level_suf_item_qid),
     Meta = list(Browser = not_applicable_qid),
     DB = list(
-        TB = questiontext_qid,
-        PTB = questiontext_qid,
-        FLB = questiontext_qid,
-        GRB = list(
-            WTXB = questiontext_qid,
-            WOTXB = questiontext_qid
-        )
+      TB = questiontext_qid,
+      PTB = questiontext_qid,
+      FLB = questiontext_qid,
+      GRB = list(
+        WTXB = questiontext_qid,
+        WOTXB = questiontext_qid
+      )
     )
   )
 
