+++ tests/testthat/_snaps/dict_validate.md
@@ -10703,47 +10703,47 @@
                 {
                   "type": "character",
                   "attributes": {},
-                  "value": ["QID10_1_1", "QID11_1_1", "QID9_1_1"]
+                  "value": ["QID10_1_1", "QID11_1_1", "QID54_1_1", "QID55_1_1", "QID9_1_1"]
                 },
                 {
                   "type": "character",
                   "attributes": {},
-                  "value": ["QID10_1_2", "QID11_1_2", "QID9_1_2"]
+                  "value": ["QID10_1_2", "QID11_1_2", "QID54_1_2", "QID55_1_2", "QID9_1_2"]
                 },
                 {
                   "type": "character",
                   "attributes": {},
-                  "value": ["QID10_1_3", "QID11_1_3", "QID9_1_3"]
+                  "value": ["QID10_1_3", "QID11_1_3", "QID54_1_3", "QID55_1_3", "QID9_1_3"]
                 },
                 {
                   "type": "character",
                   "attributes": {},
-                  "value": ["QID10_2_1", "QID11_2_1", "QID9_2_1"]
+                  "value": ["QID10_2_1", "QID11_2_1", "QID54_2_1", "QID55_2_1", "QID9_2_1"]
                 },
                 {
                   "type": "character",
                   "attributes": {},
-                  "value": ["QID10_2_2", "QID11_2_2", "QID9_2_2"]
+                  "value": ["QID10_2_2", "QID11_2_2", "QID54_2_2", "QID55_2_2", "QID9_2_2"]
                 },
                 {
                   "type": "character",
                   "attributes": {},
-                  "value": ["QID10_2_3", "QID11_2_3", "QID9_2_3"]
+                  "value": ["QID10_2_3", "QID11_2_3", "QID54_2_3", "QID55_2_3", "QID9_2_3"]
                 },
                 {
                   "type": "character",
                   "attributes": {},
-                  "value": ["QID10_3_1", "QID11_3_1", "QID9_3_1"]
+                  "value": ["QID10_3_1", "QID11_3_1", "QID54_3_1", "QID55_3_1", "QID9_3_1"]
                 },
                 {
                   "type": "character",
                   "attributes": {},
-                  "value": ["QID10_3_2", "QID11_3_2", "QID9_3_2"]
+                  "value": ["QID10_3_2", "QID11_3_2", "QID54_3_2", "QID55_3_2", "QID9_3_2"]
                 },
                 {
                   "type": "character",
                   "attributes": {},
-                  "value": ["QID10_3_3", "QID11_3_3", "QID9_3_3"]
+                  "value": ["QID10_3_3", "QID11_3_3", "QID54_3_3", "QID55_3_3", "QID9_3_3"]
                 },
                 {
                   "type": "character",
