+++ tests/fixtures/dict_generate.yml
@@ -1,5 +1,791 @@
 http_interactions:
 - request:
+    method: post
+    uri: https://QUALTRICS_BASE_URL/API/v3/surveys/SV_2l4GhidotLEBzYp/export-responses/
+    body:
+      encoding: ''
+      string: '{"format":"csv","useLabels":false,"timeZone":"Asia/Shanghai","limit":1,"includeDisplayOrder":true,"breakoutSets":true}'
+    headers:
+      X-API-TOKEN: QUALTRICS_API_KEY
+      Content-Type: application/json
+      Accept: '*/*'
+      accept-encoding: gzip, deflate
+  response:
+    status:
+      status_code: 200
+      category: Success
+      reason: OK
+      message: 'Success: (200) OK'
+    headers:
+      request-time: '1862'
+      x-frame-options: DENY
+      x-xss-protection: 1; mode=block
+      x-permitted-cross-domain-policies: master-only
+      content-type: application/json
+      content-length: '174'
+      referrer-policy: strict-origin-when-cross-origin
+      x-content-type-options: nosniff
+      permissions-policy: camera=(), geolocation=(), microphone=()
+      date: Sat, 04 Feb 2023 04:50:12 GMT
+      content-security-policy-report-only: frame-ancestors 'self' *.qualtrics.com
+        *.my.salesforce.com *.visualforce.com *.visual.force.com *.lightning.force.com;
+        report-uri https://sjc1.qualtrics.com/csp-report
+      strict-transport-security: max-age=31536000; includeSubDomains; preload
+    body:
+      encoding: ''
+      file: no
+      string: '{"result":{"progressId":"ES_2soyekGWKtNFkdo","percentComplete":0.0,"status":"inProgress"},"meta":{"requestId":"bf36b6a9-dee5-4b2d-861b-f98a50e734c4","httpStatus":"200
+        - OK"}}'
+  recorded_at: 2023-02-04 04:50:30 GMT
+  recorded_with: vcr/1.2.0, webmockr/0.8.2
+- request:
+    method: get
+    uri: https://QUALTRICS_BASE_URL/API/v3/surveys/SV_2l4GhidotLEBzYp/export-responses/ES_2soyekGWKtNFkdo
+    body:
+      encoding: ''
+      string: ''
+    headers:
+      X-API-TOKEN: QUALTRICS_API_KEY
+      Content-Type: application/json
+      Accept: '*/*'
+      accept-encoding: gzip, deflate
+  response:
+    status:
+      status_code: 200
+      category: Success
+      reason: OK
+      message: 'Success: (200) OK'
+    headers:
+      request-time: '266'
+      x-frame-options: DENY
+      x-xss-protection: 1; mode=block
+      x-permitted-cross-domain-policies: master-only
+      content-type: application/json
+      content-length: '140'
+      referrer-policy: strict-origin-when-cross-origin
[…]
