### 2023-02-11 #572 @maurolepore
Dear @lyh970817 I'm delighted to announce that I'll be the handling editor of this submission.

- [ ] I'll review the check results. If I see good opportunities to make reviewer's work easier I'll let you know.
- [ ] Can you please suggest three reviewers? Following our [guidelines](https://devguide.ropensci.org/editorguide.html#where-to-look-for-reviewers) I'll use one at most, but I would like your view of the types of expertise needed to review qualtdict.
