### 2023-02-11 #572 @lyh970817
> Dear @lyh970817 I'm delighted to announce that I'll be the handling editor of this submission.
>
> ### Semantic tags for my comments
> To help you track my comments I tagged them with "ml" and numbered sequentially: ml01, ml02, and so on. Comments following bullets are for you to consider -- you may or may not respond to them. Comments following check-boxes are requests for some action -- please respond.
>
> ### Reviewers
> * [ ]  ml01. Can you please suggest three reviewers? Following our [guidelines](https://devguide.ropensci.org/editorguide.html#where-to-look-for-reviewers) I'll use one at most, but I would like your view of the types of expertise needed to review qualtdict.
>
> ### Checks
> Here I list a few things that caught my attention. They are not blockers but the sooner we address them the better.
>
> Package Dependencies
>
> * ml02. Some imported packages appear to have no associated function calls; please ensure with author that these 'Imports' are listed appropriately.
>
> `goodpractice` and other checks
>
> * ml03. R CMD check generated the following check_fail: no_import_package_as_a_whole
> * ml04. Avoid library() and require() calls in packages: 1 time

ml01. Unfortunately I'm not sure if I could name any specific authors. But expertise-wise I thought having someone with a psychology/social science background might be helpful. As `qualtdict`  is centred around creating a variable dictionary giving an intuitive overview of survey data for analysts. The usefulness of such a dictionary is probably best judged by someone who analyses such data on a daily basis (in contrast to a data engineer who implements APIs for such data).

ml02. R CMD Check seems to fail without importing some of the packages that I don't actually use. For instance, without importing `haven`:
```r
Error in `set_labels_helper(x = .dat, labels = labels, force.labels = forc
e.labels,
    force.values = force.values, drop.na = drop.na, var.name = NULL)`: Pac
kage 'haven' required for this function. Please install it.
```
ml03. I use `dplyr`, `purrr` and `stringr` extensively so I import them as a whole. Should I still import functions from them (which will be many) individually?

ml04. I think it comes from this line in the tests:
```r
library(vcr) # *Required* as vcr is set up on loading
```
which is mandatory for `vcr` to work.
