### 2023-02-12 #572 @maurolepore
* ml02. Following your example with the haven package I saw you need to import `haven::read_xpt` because the sjlabelled package needs it. That surprises me. Usually each package must import any external function it needs, and not as users to do it. Do you know why that's the case? Also I see haven is listed in .pre-commit.config.yaml -- which I'm not familiar with. So likely there is a good explanation and I just happen to never have encounter a case like this. It would be good to articulate an explanation because reviewers might be surprised too.

* ml03. Yeah, AFAIK best practice is to either namespace each function each time you call it or import each function individually. For example, each time use something like `dplyr::filter()` or import it once  with `usethis::use_import_from("dplyr", "filter")` then use it each time just like `filter()`.

* ml04. I see. Thanks!

- [ ] ml05. When tests run I see a lot of printed output. Please suppress it so that reviewers can see a succinct test report. If the output is not generated from an R condition (e.g. messages, warnings, or errors) it may be hard to suppress. See `capture.output()` -- you may need to implement a way to capture the output and maybe implement a `quietly` argument you can set to `TRUE` during tests.

- [ ] ml06. The test results I see show many warnings. Please address them if you don't expect them or suppress them if you do expect them. If you expect them it's best to make them go away so that you don't develop the habit of ignoring them and risk missing an important one that you don't expect.

```
[ FAIL 0 | WARN 591 | SKIP 0 | PASS 4 ]
```

- [ ] Can you please make your project an RStudio project? Most R developers/contributors work in RStudio. Without an .Rproj file launching the project is hard, and I would like reviewers to enter your package as smoothly as possible. You may use `usethis::use_rstudio()`. And later it may help to lower the entry-barrier for contributors.
