### 2023-04-04 92a8b27
Remove names of `label` and `level` in `dict_validate` and update tests
