### 2024-04-08 #572 @maurolepore
@lyh970817, how would you like to proceed?

1. Follow up on https://github.com/ropensci/software-review/issues/572#issuecomment-1537062517
2. Continue on hold.
3. Withdrawal the submission.

> The holding status will be revisited every 3 months, and after one year the issue will be closed.
-- https://devdevguide.netlify.app/softwarereview_policies.html#policiesreviewprocess
