### 2024-06-19 #572 @maurolepore
Dear @lyh970817

I hope all is well. I totally understand priorities change. At this moment I believe this policy applies:

> If the author hasn’t requested a holding label, but is simply not responding, we should close the issue within one month after the last contact intent. This intent will include a comment tagging the author, but also an email using the email address listed in the DESCRIPTION of the package which is one of the rare cases where the editor will try to contact the author by email.
-- https://devdevguide.netlify.app/softwarereview_policies

FYI my next step is to confirm with the chief editor and if they agree I'll close the issue and let you know by email.
